Updated based on PR review
