+++ homeassistant/components/kaleidescape/__init__.py
@@ -2,8 +2,9 @@
 
 from __future__ import annotations
 
+from dataclasses import dataclass
 import logging
-from typing import TYPE_CHECKING, NamedTuple
+from typing import TYPE_CHECKING
 
 from kaleidescape import Device as KaleidescapeDevice, KaleidescapeError
 
@@ -21,16 +22,6 @@ _LOGGER = logging.getLogger(__name__)
 PLATFORMS = [Platform.MEDIA_PLAYER, Platform.REMOTE]
 
 
-class DeviceInfo(NamedTuple):
-    """Metadata for a Kaleidescape device."""
-
-    host: str
-    serial: str
-    name: str
-    model: str
-    server_only: bool
-
-
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up Kaleidescape from a config entry."""
     device = KaleidescapeDevice(
@@ -46,7 +37,7 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
 
     hass.data.setdefault(DOMAIN, {})[entry.entry_id] = device
 
-    async def disconnect(event: Event):
+    async def disconnect(event: Event) -> None:
         await device.disconnect()
 
     entry.async_on_unload(
@@ -66,16 +57,27 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     return unload_ok
 
 
+@dataclass
+class KaleidescapeDeviceInfo:
+    """Metadata for a Kaleidescape device."""
+
+    host: str
+    serial: str
+    name: str
+    model: str
+    server_only: bool
+
+
 class UnsupportedError(HomeAssistantError):
     """Error for unsupported device types."""
 
 
-async def validate_host(host: str) -> DeviceInfo:
+async def validate_host(host: str) -> KaleidescapeDeviceInfo:
     """Validate device host."""
     device = KaleidescapeDevice(host)
     try:
         await device.connect()
-        return DeviceInfo(
+        return KaleidescapeDeviceInfo(
             host=device.host,
             serial=device.system.serial_number,
             name=device.system.friendly_name,
