+++ homeassistant/components/kaleidescape/media_player.py
@@ -3,9 +3,7 @@
 from __future__ import annotations
 
 import logging
-from typing import TYPE_CHECKING
-
-from kaleidescape import const as kaleidescape_const
+from typing import TYPE_CHECKING, cast
 
 from homeassistant.components.media_player import MediaPlayerEntity
 from homeassistant.components.media_player.const import (
@@ -17,12 +15,10 @@ from homeassistant.components.media_player.const import (
     SUPPORT_TURN_OFF,
     SUPPORT_TURN_ON,
 )
-from homeassistant.const import STATE_IDLE, STATE_OFF, STATE_PAUSED, STATE_PLAYING
-from homeassistant.core import callback
-from homeassistant.helpers.entity import DeviceInfo
 from homeassistant.util.dt import utcnow
 
-from .const import DOMAIN as KALEIDESCAPE_DOMAIN, NAME as KALEIDESCAPE_NAME
+from .const import DOMAIN as KALEIDESCAPE_DOMAIN
+from .entity import KALEIDESCAPE_PLAYING_STATES, KaleidescapeEntity
 
 if TYPE_CHECKING:
     from datetime import datetime
@@ -44,23 +40,6 @@ SUPPORTED_FEATURES = (
     | SUPPORT_PREVIOUS_TRACK
 )
 
-KALEIDESCAPE_DEVICE_EVENTS = [
-    kaleidescape_const.DEVICE_POWER_STATE,
-    kaleidescape_const.FRIENDLY_NAME,
-    kaleidescape_const.PLAY_STATUS,
-    kaleidescape_const.MOVIE_LOCATION,
-    kaleidescape_const.SCREEN_MASK,
-    kaleidescape_const.VIDEO_COLOR,
-]
-
-KALEIDESCAPE_PLAYING_STATES = [
-    kaleidescape_const.PLAY_STATUS_PLAYING,
-    kaleidescape_const.PLAY_STATUS_FORWARD,
-    kaleidescape_const.PLAY_STATUS_REVERSE,
-]
-
-KALEIDESCAPE_PAUSED_STATES = [kaleidescape_const.PLAY_STATUS_PAUSED]
-
 _LOGGER = logging.getLogger(__name__)
 
 
@@ -69,30 +48,16 @@ async def async_setup_entry(
 ) -> None:
     """Set up the platform from a config entry."""
     entities = [KaleidescapeMediaPlayer(hass.data[KALEIDESCAPE_DOMAIN][entry.entry_id])]
-    async_add_entities(entities, True)
+    async_add_entities(entities)
 
 
-class KaleidescapeMediaPlayer(MediaPlayerEntity):
+class KaleidescapeMediaPlayer(KaleidescapeEntity, MediaPlayerEntity):
     """Representation of a Kaleidescape device."""
 
     def __init__(self, device: KaleidescapeDevice) -> None:
         """Initialize media player."""
-        self._device = device
-
-        self._attr_should_poll = False
-        self._attr_unique_id = device.serial_number
-        self._attr_name = f"{KALEIDESCAPE_NAME} {device.system.friendly_name}"
+        super().__init__(device)
         self._attr_supported_features = SUPPORTED_FEATURES
-
-    async def async_added_to_hass(self) -> None:
-        """Register update listener."""
-
-        @callback
-        def _update(event: str) -> None:
-            """Handle device state changes."""
-            self.async_write_ha_state()
-
-        self.async_on_remove(self._device.dispatcher.connect(_update).disconnect)
 
     async def async_turn_on(self) -> None:
         """Send leave standby command."""
@@ -114,89 +79,43 @@ class KaleidescapeMediaPlayer(MediaPlayerEntity):
         """Send stop command."""
         await self._device.stop()
 
-    async def async_media_next_track(self):
+    async def async_media_next_track(self) -> None:
         """Send track next command."""
         await self._device.next()
 
-    async def async_media_previous_track(self):
+    async def async_media_previous_track(self) -> None:
         """Send track previous command."""
         await self._device.previous()
 
     @property
     def available(self) -> bool:
         """Return if device is available."""
-        return self._device.is_connected
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        """Return device specific attributes."""
-        return DeviceInfo(
-            identifiers={(KALEIDESCAPE_DOMAIN, self._device.serial_number)},
-            name=self.name,
-            model=self._device.system.type,
-            manufacturer=KALEIDESCAPE_NAME,
-            sw_version=f"{self._device.system.kos_version}",
-            suggested_area="Theater",
-            configuration_url=f"http://{self._device.host}",
-        )
-
-    @property
-    def extra_state_attributes(self) -> dict:
-        """Return additional attributes about the state."""
-        return {
-            "media_location": self._device.automation.movie_location,
-            "play_status": self._device.movie.play_status,
-            "play_speed": self._device.movie.play_speed,
-            "video_mode": self._device.automation.video_mode,
-            "video_color_eotf": self._device.automation.video_color_eotf,
-            "video_color_space": self._device.automation.video_color_space,
-            "video_color_depth": self._device.automation.video_color_depth,
-            "video_color_sampling": self._device.automation.video_color_sampling,
-            "screen_mask_ratio": self._device.automation.screen_mask_ratio,
-            "screen_mask_top_trim_rel": self._device.automation.screen_mask_top_trim_rel,
-            "screen_mask_bottom_trim_rel": self._device.automation.screen_mask_bottom_trim_rel,
-            "screen_mask_conservative_ratio": self._device.automation.screen_mask_conservative_ratio,
-            "screen_mask_top_mask_abs": self._device.automation.screen_mask_top_mask_abs,
-            "screen_mask_bottom_mask_abs": self._device.automation.screen_mask_bottom_mask_abs,
-            "cinemascape_mask": self._device.automation.cinemascape_mask,
-            "cinemascape_mode": self._device.automation.cinemascape_mode,
-        }
-
-    @property
-    def state(self) -> str:
-        """State of device."""
-        if self._device.power.state == kaleidescape_const.DEVICE_POWER_STATE_STANDBY:
-            return STATE_OFF
-        if self._device.movie.play_status in KALEIDESCAPE_PLAYING_STATES:
-            return STATE_PLAYING
-        if self._device.movie.play_status in KALEIDESCAPE_PAUSED_STATES:
-            return STATE_PAUSED
-        return STATE_IDLE
+        return cast(bool, self._device.is_connected)
 
     @property
     def media_content_id(self) -> str | None:
         """Content ID of current playing media."""
         if self._device.movie.handle:
-            return self._device.movie.handle
+            return cast(str, self._device.movie.handle)
         return None
 
     @property
     def media_content_type(self) -> str | None:
         """Content type of current playing media."""
-        return self._device.movie.media_type
+        return cast(str, self._device.movie.media_type)
 
     @property
     def media_duration(self) -> int | None:
         """Duration of current playing media in seconds."""
         if self._device.movie.title_length:
-            return self._device.movie.title_length
+            return cast(int, self._device.movie.title_length)
         return None
 
     @property
     def media_position(self) -> int | None:
         """Position of current playing media in seconds."""
         if self._device.movie.title_location:
-            return self._device.movie.title_location
+            return cast(int, self._device.movie.title_location)
         return None
 
     @property
@@ -209,9 +128,9 @@ class KaleidescapeMediaPlayer(MediaPlayerEntity):
     @property
     def media_image_url(self) -> str:
         """Image url of current playing media."""
-        return self._device.movie.cover
+        return cast(str, self._device.movie.cover)
 
     @property
     def media_title(self) -> str:
         """Title of current playing media."""
-        return self._device.movie.title
+        return cast(str, self._device.movie.title)
